% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{access_to_sqlite_mdbtools}
\alias{access_to_sqlite_mdbtools}
\title{Convert Microsoft Access database to SQLite using mdbtools}
\usage{
access_to_sqlite_mdbtools(
  accdb,
  sqlite = ":memory:",
  tables = NULL,
  drop = TRUE
)
}
\arguments{
\item{accdb}{[string] Path to the Microsoft Access database file.}

\item{sqlite}{[string] Path to the SQLite database file. If \verb{:memory:},
which is the default, a temporary in-memory database will be created.}

\item{tables}{[character] Vector of table names to convert from the
Microsoft Access database to the SQLite database. If \code{NULL}, which
is the default, all tables will be converted.}

\item{drop}{[logical] Whether to drop tables in the SQLite database if
they already exist. Default is \code{TRUE}.}
}
\value{
[DBIConnection] A SQLite database connection.
}
\description{
Convert Microsoft Access database to SQLite using mdbtools
}
\note{
This function requires 'mdbtools' installed and can be found in PATH,
which means that it only works on macOS and Linux. For Windows, use
\code{\link[=access_to_sqlite_dbi]{access_to_sqlite_dbi()}} instead.
}
\keyword{internal}
