% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destep-package.R
\docType{package}
\name{destep-package}
\alias{destep}
\alias{destep-package}
\title{destep: A toolkit to convert 'DeST' models to 'EnergyPlus' models}
\description{
A toolkit to convert 'DeST' models to 'EnergyPlus' models.
}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})

Authors:
\itemize{
  \item Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})
}

}
\keyword{internal}
