% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{destep_update_name}
\alias{destep_update_name}
\title{Update NAME column in DeST tables}
\usage{
destep_update_name(dest, tables = NULL)
}
\arguments{
\item{dest}{[DBIConnection] A SQLite database connection to the DeST model.}

\item{tables}{[character] Vector of table names to update. If \code{NULL}, which
is the default, all tables will be updated.}
}
\value{
[DBIConnection] The same database connection object.
}
\description{
Update NAME column in DeST tables
}
\details{
In EnergyPlus, object names have to be unique in the scope of their belonging
class. However, in DeST, 'NAME' column is likely empty or '.' for most table.
This function can be called before actual conversion starts to update the
values in 'NAME' column and make sure the follow EnergyPlus requirements.

The updating process consists of two steps:
\enumerate{
\item Fill empty name column with a prefix if the name is empty or '.'. The
prefix is based on the table name.
\item Add suffix to the names if there are duplicated names.
}
}
\keyword{internal}
