% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dest_model}
\alias{download_dest_model}
\title{Download a typical building model from the DeST website}
\usage{
download_dest_model(building_type, location, year, dir = ".")
}
\arguments{
\item{building_type}{[character] Building type to download.}

\item{location}{[character] Location of the building.}

\item{year}{[integer] Year of the building model.}

\item{dir}{[string] Directory to save the file. Default is the current
workding directory.}
}
\value{
[character] Path to the downloaded file.
}
\description{
Download a typical building model from the DeST website
}
\details{
The downloaded \verb{7z} file will be unachieved using the \code{{archive}} package.
}
\note{
Information about all available models can be retrieved by sending a
\code{GET} request to \verb{https://svr.dest.net.cn/api/v1/all_model_names}.
}
\examples{
\dontrun{
download_typical_dest("Commerical office A", "Chongqin", 2015)
}

}
