% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read__access_to_sqlite_mdbtools}
\alias{read__access_to_sqlite_mdbtools}
\title{Convert Microsoft Access database to SQLite using mdbtools}
\usage{
read__access_to_sqlite_mdbtools(
  accdb,
  sqlite = ":memory:",
  tables = NULL,
  drop = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{accdb}{[string] Path to the DeST model file. Usually a Microsoft
Access database file with \code{.accdb} extension.}

\item{sqlite}{[string] Path to the SQLite database file. If \verb{:memory:},
which is the default, a temporary in-memory database will be created.}

\item{tables}{[character] Vector of table names to read from the DeST
model. If \code{NULL}, which is the default, all tables will be read.}

\item{drop}{[logical] Whether to drop tables in the SQLite database if
they already exist. Default is \code{TRUE}.}

\item{verbose}{[logical] Whether to print information about the conversion
process. Default is \code{FALSE}.}
}
\description{
Convert Microsoft Access database to SQLite using mdbtools
}
\note{
This function requires 'mdbtools' installed and can be found in PATH,
which means that it only works on macOS and Linux. For Windows, use
\code{\link[=read__access_to_sqlite_odbc]{read__access_to_sqlite_odbc()}} instead.
}
\keyword{internal}
