% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{access_to_sqlite_dbi}
\alias{access_to_sqlite_dbi}
\title{Convert Microsoft Access database to SQLite using DBI}
\usage{
access_to_sqlite_dbi(
  accdb,
  sqlite = ":memory:",
  tables = NULL,
  drop = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{accdb}{[string] Path to the DeST model file. Usually a Microsoft
Access database file with \code{.accdb} extension.}

\item{sqlite}{[string] Path to the SQLite database file. If \verb{:memory:},
which is the default, a temporary in-memory database will be created.}

\item{tables}{[character] Vector of table names to read from the DeST
model. If \code{NULL}, which is the default, all tables will be read.}

\item{drop}{[logical] Whether to drop tables in the SQLite database if
they already exist. Default is \code{TRUE}.}

\item{verbose}{[logical] Whether to print information about the conversion
process. Default is \code{TRUE}.}
}
\description{
Convert Microsoft Access database to SQLite using DBI
}
\note{
This only works on Windows. For macOS and Linux, use
\code{\link[=access_to_sqlite_mdbtools]{access_to_sqlite_mdbtools()}} instead.
}
\keyword{internal}
