% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dest}
\alias{read_dest}
\title{Read tables from a DeST model and convert to SQLite}
\usage{
read_dest(file, tables = NULL, sqlite = ":memory:")
}
\arguments{
\item{file}{[string] Path to the DeST model file.}

\item{tables}{[character] Vector of table names to read from the DeST
model. If \code{NULL}, which is the default, all tables will be read.}

\item{sqlite}{[string] Path to the SQLite database file. If \verb{:memory:},
which is the default, a temporary in-memory database will be created.}
}
\value{
[DBIConnection] A SQLite database connection with all specified tables..
}
\description{
Read tables from a DeST model and convert to SQLite
}
\note{
Tables with the same name will be overwritten in the SQLite database.
}
